Tested working
a few tweaks.  tested for about 90 minutes spread out around the warehouse.  Works well.  Sometimes gets on 2 patterns instead of all being the same.  Changing line 117 of foc.ino according to the comment there could help this, but I don't want to reprogram all 25 again and it could be cool for them to do different things sometimes.  It could also be nice to have the brain reboot if it hasn't been connected to any other node for a while just incase the network stack crashes.  I think this happened when I unplugged a bunch and the network had to try too hard to heal.